# CITFAC-openapi.yaml
openapi: 3.0.3
info:
  title: Città Facile
  description: Insieme di API esposte dal BFF di Città Facile (nativamente o attraverso proxy di API esistenti)
  contact:
    email: maurizio.peisino@csi.it
  version: 1.0.0
servers:
  - url: 'https://localhost'
...
paths:
...
  /api-public/otp:
    post:
      operationId: postOtp
      summary: Generate OTP
      description: Ask for OTP code
      tags:
        - Public | OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                codiceFiscale:
                  type: string
                  format: codice-fiscale
                  example: GLLGLL64B15G702I
                tipoRichiesta:
                  type: string
                  description: kind of request SMS or EMAIL
                  example: SMS
                contatto:
                  type: string
                  format: phone
                  example: 00393807129011
      responses:
        '200':
          description: The request has been fulfilled correctly
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  /api-public/otp/verify:
    get:
      operationId: getVerify
      summary: Verify OTP
      description: Verify OTP code
      tags:
        - Public | OTP
      parameters:
        - name: otp
          in: query
          description: insert OTP code
          required: true
          schema:
            type: string
            default: '12345'
        - name: cf
          in: query
          description: insert codice fiscale
          required: true
          schema:
            type: string
            default: GLLGLL64B15G702I
        - name: richiesta
          in: query
          description: insert type of request SMS or EMAIL
          required: true
          schema:
            type: string
            default: SMS
      responses:
        '200':
          description: The request has been fulfilled correctly
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /api-public/otp/channels/sms/attempts-left:
    get:
      operationId: getAttempts
      summary: Check sending attempts
      description: Check OTP attempts
      tags:
        - Public | OTP
      parameters:
        - name: cf
          in: query
          description: insert codice fiscale
          required: true
          schema:
            type: string
            default: GLLGLL64B15G702I
      responses:
        '200':
          description: The request has been fulfilled correctly
  /api/preferenze:
    get:
      operationId: getAllPreferenze
      summary: Get all preferences
      description: Reports all the specific preferences of a user for a given tenant
      tags:
        - Authenticated | Preferences
      responses:
        '200':
          description: List of services details
          content:
            application/json:
              example:
                status: 200
                type: array
                items: {}
        '400':
          $ref: '#/components/responses/400'
  '/api/preferenze/{service-id}':
    parameters:
      - $ref: '#/components/parameters/idServizio'
    post:
      operationId: addPreferenza
      summary: Add new preference
      description: Add new  preferences of a user for a given tenant
      tags:
        - Authenticated | Preferences
      responses:
        '200':
          description: Preferenza aggiunta con Successo
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  example: 2
        '400':
          $ref: '#/components/responses/400'
    delete:
      operationId: removePreferenza
      summary: Delete preference
      description: Delete a preference of a user for a given tenant
      tags:
        - Authenticated | Preferences
      responses:
        '200':
          description: Preferenza eliminata con Successo
          content:
            application/json:
              example:
                status: 200
                title: Preferenza eliminata con Successo
        '400':
          $ref: '#/components/responses/400'
  /api-public/feedback:
    post:
      operationId: addFeedback
      summary: Add new feedback
      description: Add new user feedback
      tags:
        - Public | Feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feedback'
      responses:
        '200':
          description: Feedback NOT recorded (honeypot in action)
        '201':
          description: Feedback recorded
  /api-public/feedback-moon:
    post:
      operationId: addFeedbackMoon
      summary: Add new feedback from mOOn
      description: Add new user feedback coming from mOOn module
      tags:
        - Public | Feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackMoon'
      responses:
        '200':
          description: Feedback NOT recorded (honeypot in action)
        '201':
          description: Feedback recorded
...
  /api/drafts:
    parameters:
      - $ref: '#/components/parameters/filtroAsset'
    get:
      operationId: getDrafts
      summary: Get all drafts
      description: 'Get all available drafts for single user, eventually filtered by asset'
      tags:
        - Authenticated | Drafts
      responses:
        '200':
          description: List of drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bozza'
  /api/draft:
    post:
      operationId: addDraft
      summary: Add draft
      description: Add new draft for single asset
      tags:
        - Authenticated | Drafts
      responses:
        '201':
          description: Create draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bozza'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
  '/api/draft/{draft-id}':
    parameters:
      - $ref: '#/components/parameters/idBozza'
    put:
      operationId: putDraft
      summary: Update draft
      description: Update user draft for single asset
      tags:
        - Authenticated | Drafts
      responses:
        '200':
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bozza'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
    delete:
      operationId: deleteDraft
      summary: Delete draft
      description: Delete user draft for single asset
      tags:
        - Authenticated | Drafts
      responses:
        '200':
          description: Draft is deleted
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
  '/api-admin/tenant/{tenant-id}':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getTenant
      summary: Get tenant
      description: Get single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putTenant
      summary: Update tenant
      description: Update single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
    patch:
      operationId: patchTenant
      summary: Update tenant partially
      description: Update single tenant configuration passing modified fields only
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
...
  '/api-admin/asset/{asset-id}':
    parameters:
      - $ref: '#/components/parameters/idAsset'
    get:
      operationId: getAsset
      summary: Get asset
      description: Get single asset
      tags:
        - Administration | Assets
      responses:
        '200':
          description: Asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putAsset
      summary: Update asset
      description: Update single asset
      tags:
        - Administration | Assets
      responses:
        '200':
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
...
components:
  parameters:
    idComune:
      name: tenant-id
      in: path
      required: true
      schema:
        type: integer
        default: 0
    idServizio:
      name: service-id
      in: path
      required: true
      schema:
        type: integer
...
    idAsset:
      name: asset-id
      in: path
      required: true
      schema:
        type: integer
    idBozza:
      name: draft-id
      in: path
      required: true
      schema:
        type: integer
    filtroAsset:
      in: query
      name: asset
      required: false
      schema:
        type: integer
  responses:
    '400':
      description: Bad Request. Body or parameters validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
    '403':
      description: Forbidden. User have no access to resource or tenant is not corresponding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '404':
      description: Not Found. Cannot find entity key or tenant id related to the entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '422':
      description: Unprocessable Entity. Duplication of unique fields or other integrity constrain violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
  schemas:
    Errore:
      type: object
      properties:
        status:
          type: integer
          format: int32
        title:
          type: string
          example: Messaggio di errore
        detail:
          type: array
          nullable: true
          items:
            type: string
            example: 'Nome campo: Dettaglio messaggio di errore'
...
    Bozza:
      type: object
      properties:
        id:
          type: integer
          example: 1
        idAsset:
          type: integer
          example: 1
        body:
          type: string
          description: Serialized draft JSON object structured according to asset needs
        dataCreazione:
          type: string
          format: date-time
          description: Draft creation datetime
    Comune:
      type: object
      properties:
        nome:
          type: string
          example: Novara
        regione:
          $ref: '#/components/schemas/Regione'
        tenant:
          type: string
          example: comune.pinerolo.to.it
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        homePage:
          type: string
          example: 'https://www.comune.pinerolo.to.it'
        codiceEnte:
          type: string
          description: codice IPA del comune
          example: c_g388
    ComuneAdmin:
      allOf:
        - $ref: '#/components/schemas/Comune'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            templateNotificatore:
              type: string
            logoutUri:
              type: string
            cf:
              type: string
            cms:
              type: object
              properties:
                urlServizi:
                  type: string
                urlMenu:
                  type: string
                backOffice:
                  type: string
            assets:
              type: array
              items:
                type: object
                properties:
                  tipoCredenziali:
                    type: string
                  urlBackoffice:
                    type: string
                  icon:
                    type: string
            apiManagerKey:
              type: string
            apiManagerSecret:
              type: string
    Regione:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Piemonte
        link:
          type: string
          format: uri
          description: link al portale regionale
          example: 'https://www.regione.piemonte.it/'
...
    Asset:
      type: object
      properties:
        id:
          type: integer
          example: 1
        codice:
          type: string
          example: PPAY
        nome:
          type: string
          example: PiemontePay
        descrizione:
          type: string
          example: Gestione dei pagamenti per comuni del piemonte
        urlBackoffice:
          type: string
          nullable: true
          example: 'null'
        icon:
          type: string
          nullable: true
          example: 'null'
...
    Feedback:
      type: object
      properties:
        pagePath:
          type: string
          format: uri
          description: Page url referred by feedback form
          example: 'https://dev-www-comune-pinerolo-to.portali.csi.it/servizi/iscrizione-servizi-scolastici'
        pageTitle:
          type: string
          nullable: true
          description: Page title referred by feedback form (will be showed in administation pages)
          example: Iscrizione ai servizi scolastici
        serviceID:
          type: integer
          nullable: true
          description: Service ID referred by feedback form (if any)
          example: 1
        rating:
          type: integer
          description: Rating given by user
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          example: 4
        option:
          type: string
          description: Feedback option according to AgID definitions (codified)
          enum:
            - positivo1
            - positivo2
            - positivo3
            - positivo4
            - negativo1
            - negativo2
            - negativo3
            - negativo4
            - altro
        details:
          type: string
          description: Feedback details given by user in a textarea
        bind:
          type: string
          nullable: true
          description: 'Should be null or empty, any different value avoid feedback registration. this field can be used as honeypot for bots'
          example: sono un bot
    FeedbackMoon:
      title: Feedback (mOOn)
      type: object
      properties:
        subject:
          type: string
          nullable: true
          description: Oggetto del form mOOn
          example: Iscrizione ai servizi scolastici
        code:
          type: string
          nullable: true
          description: Codice modulo mOOn
          example: CODICE_MODULO
        rating:
          type: integer
          description: Rating given by user
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          example: 4
        option:
          type: string
          description: Feedback option according to AgID definitions (codified)
          enum:
            - positivo1
            - positivo2
            - positivo3
            - positivo4
            - negativo1
            - negativo2
            - negativo3
            - negativo4
            - altro
        details:
          type: string
          description: Feedback details given by user in a textarea
        bind:
          type: string
          nullable: true
          description: 'Should be null or empty, any different value avoid feedback registration. this field can be used as honeypot for bots'
          example: sono un bot
...